Review feedback
- Move some stuff into impls mod
